Read the Geographic Attributes File

Reads the Geographies Attributes File for the given census year. The table contains the information on how all the geographic levels are related for each area, and population, dwelling and household counts. A reference guide is available at https://www150.statcan.gc.ca/n1/en/catalogue/92-151-G2021001

get_geographic_attributes(census_year = "2016", refresh = FALSE)

Arguments

census_year

census year to get the data for, right now only 2006, 2011 and 2016 are supported

refresh

(logical) refresh the cache if true

Value

tibble with the relationship data

Examples

# list add the cached census data
get_geographic_attributes("2016")
#> # A tibble: 489,905 × 48
#>    `DBuid/IDidu` `DBpop2016/IDpop2016` `DBtdwell2016/IDtlog… `DBurdwell2016/IDr…
#>    <chr>                         <dbl>                 <dbl>               <dbl>
#>  1 10020117004                      36                    13                  10
#>  2 10020117003                      35                    21                  16
#>  3 10020117008                      46                    28                  21
#>  4 10010746008                      30                    28                  15
#>  5 10010746010                      26                    44                  17
#>  6 10020081027                      78                    35                  33
#>  7 10020081035                      93                    39                  37
#>  8 10020105005                       0                     0                   0
#>  9 10020105004                     100                    46                  41
#> 10 10020105006                       0                     0                   0
#> # … with 489,895 more rows, and 44 more variables: DBarea2016/IDsup2016 <dbl>,
#> #   DBir2016/IDri2016 <chr>, DAuid/ADidu <chr>, DArplamx/ADlamx <chr>,
#> #   DArplamy/ADlamy <chr>, DArplat/Adlat <chr>, DArplong/ADlong <chr>,
#> #   PRuid/PRidu <chr>, PRname/PRnom <chr>, PRename/PRanom <chr>,
#> #   PRfname/PRfnom <chr>, PReabbr/PRaabbrev <chr>, PRfabbr/PRfabbrev <chr>,
#> #   FEDuid/CEFidu <chr>, FEDname/CEFnom <chr>, ERuid/REidu <chr>,
#> #   ERname/REnom <chr>, CDuid/DRidu <chr>, CDname/DRnom <chr>, …