List all child variables from vector hierarchies given either a list of Census variables returned by list_census_vectors, search_census_vectors, find_census_vectors, or a direct string reference to the vector code.

List all child variables from vector hierarchies given either a list of Census variables returned by list_census_vectors, search_census_vectors, find_census_vectors, or a direct string reference to the vector code.

child_census_vectors(vector_list, leaves_only = FALSE, max_level = NA)

Arguments

vector_list
The list of vectors to be used, either a character vector or a filtered tibble as returned from list_census_vectors.

leaves_only
Boolean flag to indicate if only leaf vectors should be returned, i.e. vectors that don't have children.

max_level
optional, maximum depth to look for child vectors. Default is NA will return all child census vectors.

Examples

# Query parent vectors directly using vector identifier
child_census_vectors("v_CA16_2510")

#> # A tibble: 4 x 7
#>   vector   type  label    units parent_vector aggregation details               
#>   <chr>    <fct> <chr>    <fct> <chr>         <chr>       <chr>                 
#> 1 v_CA16_… Total 0 to 17… Numb… v_CA16_2510   Additive    CA 2016 Census; 100% …
#> 2 v_CA16_… Total 18 to 6… Numb… v_CA16_2510   Additive    CA 2016 Census; 100% …
#> 3 v_CA16_… Total 65 year… Numb… v_CA16_2510   Additive    CA 2016 Census; 100% …
#> 4 v_CA16_… Total 0 to 5 … Numb… v_CA16_2513   Additive    CA 2016 Census; 100% …

if (FALSE) {

# Example using multiple vectors coerced into a list
child_census_vectors(c("v_CA16_2510","v_CA16_2511","v_CA16_2512"))

# or, equivalently
selected_vectors <- c("v_CA16_2510","v_CA16_2511","v_CA16_2512")
child_census_vectors(selected_vectors)

# Example using dplyr and piped arguments
library(dplyr, warn.conflicts = FALSE)

list_census_vectors("CA16") %>%
  filter(vector == "v_CA16_2510") %>%
  child_census_vectors(TRUE)
}