Query the CensusMapper API for available datasets.

Query the CensusMapper API for available datasets.

list_census_datasets(use_cache = FALSE, quiet = FALSE)

Arguments

use_cache
If set to TRUE, data will be read from a local cache, if available. If set to FALSE (the default), query the API for the data, and refresh the local cache with the result.

quiet
When TRUE, suppress messages and warnings.

Value

Returns a data frame with a column dataset containing the code for the dataset, and a column description describing it.

Examples

# List available datasets in CensusMapper
list_census_datasets()

#> Querying CensusMapper API for available datasets...

#> # A tibble: 3 x 2
#>   dataset description               
#> * <chr>   <chr>                     
#> 1 CA06    2006 Canada Census        
#> 2 CA11    2011 Canada Census and NHS
#> 3 CA16    2016 Canada Census