Query the CensusMapper API for available datasets.

Query the CensusMapper API for available datasets.

list_census_datasets(use_cache = TRUE, quiet = FALSE)

Arguments

use_cache

If set to TRUE (the default), data will be read from a temporary local cache for the duration of the R session, if available. If set to FALSE, query the API for the data, and refresh the temporary cache with the result.

quiet

When TRUE, suppress messages and warnings.

Value

Returns a data frame with a column dataset containing the code for the dataset, a column description describing it, a geo_dataset column identifying the geography dataset the data is based on, a attribution column with an attribution string, a reference column with a reference identifier, and a reference_url column with a link to reference materials.

Examples

# List available datasets in CensusMapper
list_census_datasets()
#> Reading dataset list from temporary cache.
#> # A tibble: 29 × 6
#>    dataset description           geo_dataset attribution reference reference_url
#>    <chr>   <chr>                 <chr>       <chr>       <chr>     <chr>        
#>  1 CA1996  1996 Canada Census    CA1996      StatCan 19… 92-351-U  https://www1…
#>  2 CA01    2001 Canada Census    CA01        StatCan 20… 92-378-X  https://www1…
#>  3 CA06    2006 Canada Census    CA06        StatCan 20… 92-566-X  https://www1…
#>  4 CA11    2011 Canada Census a… CA11        StatCan 20… 98-301-X… https://www1…
#>  5 CA16    2016 Canada Census    CA16        StatCan 20… 98-301-X  https://www1…
#>  6 CA21    2021 Canada Census    CA21        StatCan 20… 98-301-X  https://www1…
#>  7 CA01xSD 2001 Canada Census x… CA01        StatCan 20… 92-378-X  https://www1…
#>  8 CA06xSD 2006 Canada Census x… CA06        StatCan 20… 92-566-X  https://www1…
#>  9 CA11xSD 2011 Canada Census x… CA11        StatCan 20… 98-301-X  https://www1…
#> 10 CA16xSD 2016 Canada Census x… CA16        StatCan 20… 98-301-X  https://www1…
#> # ℹ 19 more rows