Query the CensusMapper API for vectors with descriptions matching a searchterm.

Query the CensusMapper API for vectors with descriptions matching a searchterm.

search_census_vectors(searchterm, dataset, type = NA, ...)

Arguments

searchterm
The term to search for e.g. "Ojibway". Search terms are case insensitive. If unable to find a given search term, this function will suggest the correct spelling to use when possible.

dataset
The dataset to query for available vectors, e.g. "CA16".

type
One of NA, 'Total', 'Male' or 'Female'. If specified, only return variables of specified `type`.

...
Further arguments passed on to list_census_vectors.

Examples

search_census_vectors('Ojibway', 'CA16')

#> # A tibble: 24 x 7
#>         vector   type                        label  units parent_vector
#>          <chr> <fctr>                        <chr> <fctr>         <chr>
#>  1  v_CA16_617  Total Ojibway-Potawatomi languages Number    v_CA16_569
#>  2  v_CA16_618   Male Ojibway-Potawatomi languages Number    v_CA16_570
#>  3  v_CA16_619 Female Ojibway-Potawatomi languages Number    v_CA16_571
#>  4  v_CA16_623  Total                      Ojibway Number    v_CA16_617
#>  5  v_CA16_624   Male                      Ojibway Number    v_CA16_618
#>  6  v_CA16_625 Female                      Ojibway Number    v_CA16_619
#>  7 v_CA16_1424  Total Ojibway-Potawatomi languages Number   v_CA16_1376
#>  8 v_CA16_1425   Male Ojibway-Potawatomi languages Number   v_CA16_1377
#>  9 v_CA16_1426 Female Ojibway-Potawatomi languages Number   v_CA16_1378
#> 10 v_CA16_1430  Total                      Ojibway Number   v_CA16_1424
#> # ... with 14 more rows, and 2 more variables: aggregation <chr>, details <chr>
# NOT RUN {
# This will return a warning that no match was found, but will suggest similar terms.
search_census_vectors('Ojibwe', 'CA16', 'Total')
# }